Fixed compilation issues caused by setting all files in Library.Desktop to default to createing their own precompiled headers. pch.cpp now creates its precompiled header and all other files simply use it.

+++ source/UnitTest.Library.Desktop/OrderedMapTest.cpp
@@ -443,7 +443,7 @@ namespace UnitTestLibraryDesktop
 
 				removeOrder.Shuffle();
 
-				int removeOrderIndex = 0;
+				size_t removeOrderIndex = 0;
 				while (removeOrderIndex < removeOrder.Size() - 1)
 				{
 					Assert::IsTrue(orderedMapCopy == orderedMap);
