+++ source/UnitTest.Library.Desktop/OrderedMultiMapTest.cpp
@@ -464,7 +464,7 @@ namespace UnitTestLibraryDesktop
 
 				removeOrder.Shuffle();
 
-				int removeOrderIndex = 0;
+				size_t removeOrderIndex = 0;
 				while (removeOrderIndex < removeOrder.Size() - 1)
 				{
 					Assert::IsTrue(orderedMultiMapCopy == orderedMultiMap);
