+++ source/UnitTest.Library.Desktop/OrderedMultiSetTest.cpp
@@ -463,7 +463,7 @@ namespace UnitTestLibraryDesktop
 
 				removeOrder.Shuffle();
 
-				int removeOrderIndex = 0;
+				size_t removeOrderIndex = 0;
 				while (removeOrderIndex < removeOrder.Size() - 1)
 				{
 					Assert::IsTrue(orderedMultiSetCopy == orderedMultiSet);
