+++ source/UnitTest.Library.Desktop/OrderedSetTest.cpp
@@ -440,7 +440,7 @@ namespace UnitTestLibraryDesktop
 
 				removeOrder.Shuffle();
 
-				int removeOrderIndex = 0;
+				size_t removeOrderIndex = 0;
 				while (removeOrderIndex < removeOrder.Size() - 1)
 				{
 					Assert::IsTrue(orderedSetCopy == orderedSet);
